use parsingContext for better deduction when parsing invalid/incomplete code
often times incomplete expression statement would be taken as beginning of a function declaration. which in turn had an impact at some other features, such as placeholding arguments of a function call provided from code completions. in such case the method would not expand arguments with assumption that it was dealing with new function declaration, rathen than a call.